Update 2020-04-01-template.md
<!--#### Weakness #1: Poor Performance Using Local SGD Only

Looking back to the experiment A4, which only uses the local SGD throughout the training, its performance can only be described as being mediocre. The state-of-the-art performance can only be achieved whithout using mini-batch SGD to warm up the training a little bit, hence formulating the procedure of post-local SGD. Due to this, it casts some doubt into the effectiveness of the local SGD method since it seems that using it alone is not sufficient. The authors did not provide any insight into this issue.-->